% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVI_PG.R
\name{CVI_PG}
\alias{CVI_PG}
\title{Title Collapsed Variational Inference (CVI) for Poisson-Gamma model}
\usage{
CVI_PG(N, D, T0, s1, s2, A0, B0, X, W1, W2, A, B, Plog, maxit)
}
\arguments{
\item{N}{number of observed data}

\item{D}{dimension of each observed data}

\item{T0}{total clusters fixed for the variational distribution}

\item{s1}{shape parameter for the prior distribution of alpha}

\item{s2}{rate parameter for the prior distribution of alpha}

\item{A0}{shape parameter matrix for the prior distribution of lambda}

\item{B0}{rate parameter matrix for the prior distribution of lambda}

\item{X}{observed data, N x D matrix}

\item{W1}{shape parameter for the posterior distribution of alpha}

\item{W2}{rate parameter for the posterior distribution of alpha}

\item{A}{shape parameter matrix for the posterior distribution of lambda}

\item{B}{rate parameter matrix for the posterior distribution of lambda}

\item{Plog}{logarithm of latent allocation probability matrix}

\item{maxit}{maximum number of iterations the updates are run}
}
\value{
the posterior value of alpha, the number of clusters from the
latent allocation probability, the proportions of each cluster and the
latent allocation probability matrix
}
\description{
Title Collapsed Variational Inference (CVI) for Poisson-Gamma model
}
\examples{
CVI_PG(N = 100, D = 2, T0 = 10, s1 = 0.01, s2 = 0.01,
             A0 = matrix(0.001, nrow = 10, ncol = 2),
             B0 = matrix(0.001, nrow = 10, ncol = 2),
             X = matrix(c(rep(0, 50), rep(10, 50), rep(0, 50), rep(10, 50)),
                 nrow = 100, ncol = 2),
             W1 = 0.01, W2 = 0.01, A = matrix(0.001, nrow = 10, ncol = 2),
             B = matrix(0.001, nrow = 10, ncol = 2),
             Plog = matrix(-3, nrow = 100, ncol = 10), maxit = 1000)
}
